requirements.txt
# Core Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Configuration & Environment
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1

# Development & Testing (Optional)
pytest==7.4.3
httpx==0.25.2

# AI/ML Libraries
openai==1.58.1
numpy==2.3.1

# CrewAI for multi-agent system (compatible with Python 3.13)
crewai==0.148.0
crewai-tools==0.17.0

# Database & ORM
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Task Queue
celery==5.4.0
redis==5.2.1

# Additional tools
beautifulsoup4==4.12.3
requests==2.32.3
tweepy==4.14.0